API and AUTH

+++ PangWeb.Shared/Models/Product.cs
@@ -15,8 +15,8 @@
             ProductType = category;
         }
 
-        public int Id { get; set; }
-        public int ProductCategoryId { get; set; }
+        public long Id { get; set; }
+        public long ProductCategoryId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string ImgUrl { get; set; }
@@ -28,7 +28,7 @@
 
     public class ProductCategory
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Category { get; set; }
     }
 }
